use f64 values

+++ src/fns.rs
@@ -1,4 +1,4 @@
-use std::f32::consts::PI;
+use std::f64::consts::PI;
 
 use crate::types::Transformer;
 
@@ -30,66 +30,66 @@ impl PositionFn {
     }
 }
 
-fn linear_fn(t: f32, _inverted: bool) -> f32 {
+fn linear_fn(t: f64, _inverted: bool) -> f64 {
     t
 }
 
-fn quadratic_fn(t: f32, inverted: bool) -> f32 {
+fn quadratic_fn(t: f64, inverted: bool) -> f64 {
     if inverted {
-        1f32 - (1f32 - t).powi(2)
+        1f64 - (1f64 - t).powi(2)
     } else {
         t.powi(2)
     }
 }
 
-fn cubic_fn(t: f32, inverted: bool) -> f32 {
+fn cubic_fn(t: f64, inverted: bool) -> f64 {
     if inverted {
-        1f32 - (1f32 - t).powi(3)
+        1f64 - (1f64 - t).powi(3)
     } else {
         t.powi(3)
     }
 }
 
-fn quartic_fn(t: f32, inverted: bool) -> f32 {
+fn quartic_fn(t: f64, inverted: bool) -> f64 {
     if inverted {
-        1f32 - (1f32 - t).powi(4)
+        1f64 - (1f64 - t).powi(4)
     } else {
         t.powi(4)
     }
 }
 
-fn quintic_fn(t: f32, inverted: bool) -> f32 {
+fn quintic_fn(t: f64, inverted: bool) -> f64 {
     if inverted {
-        1f32 - (1f32 - t).powi(5)
+        1f64 - (1f64 - t).powi(5)
     } else {
         t.powi(5)
     }
 }
 
-fn sinusoidal_fn(t: f32, inverted: bool) -> f32 {
+fn sinusoidal_fn(t: f64, inverted: bool) -> f64 {
     if inverted {
-        1f32 - ((1f32 - t) * PI / 2f32).sin()
+        1f64 - ((1f64 - t) * PI / 2f64).sin()
     } else {
-        (t * PI / 2f32).sin()
+        (t * PI / 2f64).sin()
     }
 }
 
-fn asinusoidal_fn(t: f32, inverted: bool) -> f32 {
+fn asinusoidal_fn(t: f64, inverted: bool) -> f64 {
     if inverted {
-        1f32 - (1f32 - t).asin() / (PI / 2f32)
+        1f64 - (1f64 - t).asin() / (PI / 2f64)
     } else {
-        t.asin() / (PI / 2f32)
+        t.asin() / (PI / 2f64)
     }
 }
 
-fn arc_fn(t: f32, inverted: bool) -> f32 {
+fn arc_fn(t: f64, inverted: bool) -> f64 {
     if inverted {
-        (1f32 - (1f32 - t).powi(2)).sqrt()
+        (1f64 - (1f64 - t).powi(2)).sqrt()
     } else {
-        1f32 - (1f32 - t).sqrt()
+        1f64 - (1f64 - t).sqrt()
     }
 }
 
-fn smooth_step_fn(t: f32, _inverted: bool) -> f32 {
-    t.powi(2) * (3f32 - 2f32 * t)
+fn smooth_step_fn(t: f64, _inverted: bool) -> f64 {
+    t.powi(2) * (3f64 - 2f64 * t)
 }
